+++ src/App.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { BrowserRouter as Router, Link, Route } from "react-router-dom";
 import styled from "styled-components";
 import LoginForm from "./components/LoginForm";
-import RegistrationForm from "./components/RegistrationForm/RegistrationForm";
+import RegistrationForm from "./components/RegistrationForm";
 
 const Heading = styled.h1.attrs({
   className: "title is-3"
